[upd] Example 004 Div, changed float to double

+++ Example004_Div/Program.cs
@@ -1,4 +1,4 @@
-﻿float num1 = 31;
-float num2 = 5;
-float res = num1 / num2;
+﻿double num1 = 31;
+double num2 = 5;
+double res = num1 / num2;
 Console.WriteLine(res);
